orders $ order_products table added

+++ database/migrations/2019_09_03_220058_create_order_products_table.php
@@ -14,9 +14,9 @@ class CreateOrderProductsTable extends Migration
     public function up()
     {
         Schema::create('order_products', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('order_id');
-            $table->unsignedBigInteger('product_id');
+            $table->increments('id');
+            $table->unsignedInteger('order_id');
+            $table->unsignedInteger('product_id');
             $table->integer('quantity');
             $table->decimal('price', 10, 2);
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
